FIX: Protect against internal throws
The message regarding uncaught rejects blindly displays the error stack. In cases
where this stack is not present an error is thrown up to the soon loop and exits
the loop prematurely. Try/catch surrounding the soon function protects against this.
Additionally, do not blindly report stack on uncaught rejections.